+++ src/generator/index.js
@@ -1,6 +1,5 @@
 module.exports = {
     match: require('./match'),
-    spotlight: require('./spotlight'),
     line: require('./line'),
     lineContent: require('./line-content'),
     newLine: require('./new-line')