+++ src/generator/new-line.js
@@ -1,13 +1,13 @@
 const { newLineLength } = require('./utils');
 
-module.exports = (source, addRange) => {
+module.exports = (source, createRange) => {
     let line = 1;
 
     for (let i = 0; i < source.length; i++) {
         const nl = newLineLength(source, i);
 
         if (nl !== 0) {
-            addRange('new-line', i, i + 1, line++);
+            createRange(i, i + nl, line++);
             i += nl - 1;
         }
     }