+++ src/generator/spotlight.js
@@ -1,4 +1,0 @@
-module.exports =
-    (...ranges) =>
-        (source, addRange) =>
-            ranges.forEach(range => addRange('spotlight', ...range));