+++ test/helpers/generators.js
@@ -1,7 +1,0 @@
-const stubGeneratorFactory =
-    ranges => (source, createRange) =>
-        ranges.forEach(range => createRange(...range));
-
-module.exports = {
-    stub: stubGeneratorFactory
-};